# docs/swagger.yaml
definitions:
  errcode.Error:
    type: object
  service.User:
    properties:
      description:
        type: string
      email:
        type: string
      phone_number:
        type: string
      sex:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a passwd server, for saving platform password.
  termsOfService: http://swagger.io/terms/
  title: passwd API
  version: "1.0"
paths:
  /api/v1/user:
    get:
      description: 获取用户分页
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.User'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个用户
      tags:
      - User
    post:
      consumes:
      - application/json
      description: 通过一些字段创建用户
      parameters:
      - description: 用户名
        in: body
        name: user_name
        required: true
        schema:
          type: string
      - description: 用户密码
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 手机号码
        in: body
        name: phone_number
        required: true
        schema:
          type: string
      - description: 电子邮件
        in: body
        name: email
        schema:
          type: string
      - description: 性别
        in: body
        name: sex
        schema:
          type: integer
      - description: 用户简介
        in: body
        name: description
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 创建用户
      tags:
      - User
  /api/v1/user/{id}:
    delete:
      description: 通过user_id删除单个用户
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除单个用户
      tags:
      - User
    get:
      consumes:
      - application/json
      description: 通过user_id获取单个文章
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.User'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取单个文章
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 通过user_id和一些model字段更新单个用户
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户名
        in: body
        name: user_name
        required: true
        schema:
          type: string
      - description: 用户密码
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 手机号码
        in: body
        name: phone_number
        required: true
        schema:
          type: string
      - description: 电子邮件
        in: body
        name: email
        schema:
          type: string
      - description: 性别
        in: body
        name: sex
        schema:
          type: integer
      - description: 用户简介
        in: body
        name: description
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新单个用户
      tags:
      - User
swagger: "2.0"
tags:
- name: restful api